feat: 🎸 update polkadot

+++ web3/polkadot-sdk-solo-template-dev-courses/pallets/kitties/src/events.rs
@@ -3,11 +3,13 @@ use frame_support::pallet_macros::pallet_section;
 /// Define all events used in the pallet.
 #[pallet_section]
 mod events {
+
     #[pallet::event]
+    #[pallet::generate_deposit(pub(super) fn deposit_event)]
     pub enum Event<T: Config> {
         KittyCreated {
             creator: T::AccountId,
-            index: u64,
+            index: u32,
             data: [u8; 16],
         },
     }
